add : components(Aside,Comment,CommentList,Feeds,Header
