+++ src/pages/chan/Login/Login.scss
@@ -29,7 +29,7 @@
   }
 
   button {
-    background: #bfdffd;
+    background: #064b8c;
     border: none;
     border-radius: 10px;
     color: #ffffff;
@@ -40,6 +40,10 @@
     margin-bottom: 100px;
     padding: 10px;
     width: 100%;
+  }
+
+  main form [type='submit']:disabled {
+    background-color: #cbddf7;
   }
 
   * {
