+++ src/pages/chan/Main/Main.scss
@@ -5,80 +5,10 @@
   height: 100%;
 }
 
-.header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 100;
-  height: 55px;
-  background: #fff;
-  border-bottom: 1px solid #dbdbdb;
-}
-
-.header .inner {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  height: 100%;
-}
-.header .inner .nav .link_list li a {
-  display: block;
-  margin: 5px;
-}
-
-.header .inner .nav .link_list li a i {
-  color: #000;
-  font-size: 30px;
-}
-
-.header .inner .logo a img {
-  width: 100px;
-  display: block;
-}
-.header .inner .input_box {
-  width: 200px;
-  height: 30px;
-}
-.header .inner .input_box input {
-  height: 100%;
-  width: 100%;
-  padding: 0 32px;
-  background: #f1f3f5;
-  border: 1px solid #ced4da;
-  font-size: 13px;
-  outline: none;
-  background: none;
-  border-radius: 5px;
-}
-
-.header .inner .input_box input::placeholder {
-  text-align: center;
-  font-size: 13px;
-}
-.header .inner .input_box input:focus::-webkit-input-placeholder {
-  color: transparent;
-}
-.header .inner .nav {
-}
-.header .inner .nav .link_list {
-  list-style: none;
-  display: flex;
-  align-items: center;
-}
-
-.header .inner .nav .link_list a img {
-  width: 50%;
-}
-/* //header */
-
 .Contents {
-  // position: relative;
-  // top: 50px;
   margin-top: 50px;
   padding: 20px;
   height: 100vh;
-  // border: solid 2px rgb(201, 28, 28);
   background: #f9f8f8;
   display: flex;
   justify-content: center;
@@ -90,276 +20,12 @@
   z-index: 1;
   top: 10px;
   left: 80px;
-  // border: solid 2px rgb(201, 28, 28);
-  // height: auto;
   width: 700px;
-  // overflow: hidden;
 }
 
 .article {
   z-index: 1;
   border: solid 2px rgb(220, 218, 221);
-  // height: auto;
   width: 540px;
   margin: 10px auto;
-  // overflow: hidden;
-}
-
-.article .feed_profile .profile_image {
-  margin: 10px;
-  font-weight: bold;
-  width: 40px;
-  height: 40px;
-  overflow: hidden;
-  display: block;
-  border-radius: 50%;
-}
-
-.feed_profile {
-  display: flex;
-}
-
-.profile_id {
-  //border: solid 1px black;
-  width: 100%;
-  height: 30px;
-  margin: 20px auto 0px auto;
-  font-size: 20px;
-  font-weight: bold;
-  padding-left: 10px;
-}
-
-.feed_id_add {
-  background-color: transparent !important;
-  background-image: none !important;
-  border-color: transparent;
-  border: none;
-
-  color: #ffffff;
-}
-
-.feed_photo {
-  /* border: solid 1px black; */
-  width: 100%;
-  height: 540px;
-  margin: 10px auto;
-}
-
-.feed_photo img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-article .article-bottom {
-  width: 100%;
-}
-
-article .article-bottom .article-links {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  padding: 10px;
-}
-
-article .article-bottom .article-links-left i {
-  font-size: 25px;
-  font-weight: normal;
-  width: 25px;
-  height: 25px;
-  margin-right: 10px;
-}
-
-// article .article-bottom .article-links-left i:last-child {
-//   margin-right: 0;
-// }
-
-article .article-bottom .article-links-right i {
-  font-size: 25px;
-  font-weight: normal;
-  width: 25px;
-  height: 25px;
-}
-
-article .article-bottom .article-bottom-like {
-  width: 100%;
-  padding: 10px;
-}
-
-article .article-bottom .article-bottom-like .like-profile .article-bold {
-  font-weight: bold;
-}
-
-.CommentList {
-  // z-index: ;
-  padding: 10px;
-  width: 100%;
-}
-
-.CommentLeft {
-  display: flex;
-}
-
-.userCommentBox {
-  display: flex;
-  justify-content: space-between;
-
-  margin-bottom: 5px;
-}
-.userName {
-  font-weight: bold;
-  margin-right: 10px;
-}
-
-.CommentList .article-bold {
-  font-weight: bold;
-  margin-right: 10px;
-}
-
-.CommentList .upload-time span {
-  color: #9a9a9a;
-}
-
-// .comment-text {
-//   margin-top: 10px;
-
-.like-profile .like_profile_image {
-  margin-right: 10px;
-  font-weight: bold;
-  width: 40px;
-  height: 40px;
-  overflow: hidden;
-  display: block;
-  border-radius: 50%;
-}
-
-.like-profile {
-  display: flex;
-  align-items: center;
-}
-
-.Comment {
-  width: 540px;
-  padding: 10px;
-  // border: 2px solid #dbdbdb;
-  margin: auto;
-}
-
-.Comment form {
-  display: flex;
-  align-items: center;
-}
-
-.Comment form .emotion {
-  width: 5%;
-  height: 20px;
-}
-
-.Comment form .emotion i {
-  font-size: 20px;
-}
-
-.Comment form input {
-  width: 85%;
-  height: 30px;
-  outline: none;
-  border: none;
-  resize: none;
-  line-height: 25px;
-  text-indent: 5px;
-}
-
-.Comment form button {
-  width: 10%;
-  height: 30px;
-  outline: none;
-  border: none;
-  background-color: #0095f6;
-}
-
-.Comment form [type='submit']:disabled {
-  background-color: #cbddf7;
-}
-
-aside {
-  // position: absolute;
-  z-index: 4;
-  top: 10px;
-  right: 80px;
-  //   border: solid 2px rgb(201, 28, 28);
-  height: 945px;
-  width: 700px;
-  // margin: auto;
-}
-
-.side-profile {
-  //   border: solid 2px rgb(201, 28, 28);
-  height: 100px;
-  width: 500px;
-  margin: 10px auto;
-  display: flex;
-  align-items: center;
-  font-size: 30px;
-  font-weight: bold;
-  padding-left: 10px;
-}
-.story-box {
-  border: 1px solid #ced4da;
-  height: 300px;
-  width: 500px;
-  background-color: #ffffff;
-  margin: 10px auto;
-  font-size: 20px;
-  font-weight: bold;
-  padding-left: 10px;
-  padding-top: 15px;
-  //paddi
-  overflow: auto;
-}
-.recommand-box {
-  border: 1px solid #ced4da;
-  height: 300px;
-  width: 500px;
-  background-color: #ffffff;
-  margin: 10px auto;
-  font-size: 20px;
-  font-weight: bold;
-  padding-left: 10px;
-  padding-top: 15px;
-  overflow: auto;
-}
-
-.side_profile_image {
-  margin: 10px;
-  font-weight: bold;
-  width: 80px;
-  height: 80px;
-  overflow: hidden;
-  display: block;
-  border-radius: 50%;
-}
-
-.story_person_image {
-  margin: 10px;
-  font-weight: bold;
-  width: 50px;
-  height: 50px;
-  overflow: hidden;
-  display: block;
-  border-radius: 50%;
-}
-
-.story-person {
-  display: flex;
-  align-items: center;
-}
-.last-box {
-  //   border: 1px solid #ced4da;
-  height: 100px;
-  width: 500px;
-  //   background-color: #FFFFFF;
-  margin: 20px auto;
-  font-size: 15px;
-  //   font-weight: bold;
-  padding-left: 10px;
-  text-align: center;
 }
