+++ src/pages/chan/components/Aside/Aside.scss
@@ -1,0 +1,75 @@
+aside {
+  z-index: 4;
+  top: 10px;
+  right: 80px;
+  height: 945px;
+  width: 700px;
+}
+
+.side-profile {
+  height: 100px;
+  width: 500px;
+  margin: 10px auto;
+  display: flex;
+  align-items: center;
+  font-size: 30px;
+  font-weight: bold;
+  padding-left: 10px;
+}
+.story-box {
+  border: 1px solid #ced4da;
+  height: 300px;
+  width: 500px;
+  background-color: #ffffff;
+  margin: 10px auto;
+  font-size: 20px;
+  font-weight: bold;
+  padding-left: 10px;
+  padding-top: 15px;
+  overflow: auto;
+}
+.recommand-box {
+  border: 1px solid #ced4da;
+  height: 300px;
+  width: 500px;
+  background-color: #ffffff;
+  margin: 10px auto;
+  font-size: 20px;
+  font-weight: bold;
+  padding-left: 10px;
+  padding-top: 15px;
+  overflow: auto;
+}
+
+.side_profile_image {
+  margin: 10px;
+  font-weight: bold;
+  width: 80px;
+  height: 80px;
+  overflow: hidden;
+  display: block;
+  border-radius: 50%;
+}
+
+.story_person_image {
+  margin: 10px;
+  font-weight: bold;
+  width: 50px;
+  height: 50px;
+  overflow: hidden;
+  display: block;
+  border-radius: 50%;
+}
+
+.story-person {
+  display: flex;
+  align-items: center;
+}
+.last-box {
+  height: 100px;
+  width: 500px;
+  margin: 20px auto;
+  font-size: 15px;
+  padding-left: 10px;
+  text-align: center;
+}
